// frontend/src/types.ts
export interface ReleasedMovie {
  id: string;
  title: string;
  slug: string;
  description: string;
  posterUrl: string;
  ticketPrice: number;
  ticketsSold: number;
  totalRevenue: number;
  author: string;
  ticketStatus: 'Draft' | 'Live';
  watchUrl?: string;
}